requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
Django==4.2.18
django-crispy-forms==2.3
djangorestframework==3.15.2
idna==3.10
newsapi-python==0.2.7
nose==1.3.7
pillow==11.1.0
requests==2.32.3
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==2.3.0